Add Post class and routing
A post class was made, containing a title, a path for the video file, and a description.
Besides that, in the post class script, a postStorage list of posts was added while the
database is not implemented. For the PostRouting script, routing for http requests (GET, POST, DELETE)
is currently being added.